requirements.txt
cachelib==0.1.1
click==7.1.2
distro==1.5.0
Flask==1.1.2
Flask-Mail==0.9.1
Flask-Session==0.3.2
fpdf==1.7.2
gcloud==0.18.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
oauth2client==4.1.3
opencv-python==4.4.0.42
packaging==20.4
Pillow==7.2.0
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytesseract==0.3.6
rsa==4.6
six==1.15.0
Werkzeug==1.0.1
